pyproject.toml
[project]
name = "pandora"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "", email = ""},
]
dependencies = ["discord-py>=2.4.0", "pytz>=2025.1", "toml>=0.10.2", "pandas>=2.2.3"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
